打成jar包时，将源文件也打入jar包

+++ java/jar/FileManagerUI.bat
@@ -12,11 +12,27 @@ if "%num%" geq "0" if "%num%" leq "9" (
 )
  
 if "%candp%"=="1" (
-    "%JAVA_HOME%javac.exe" -encoding UTF-8 -cp "../lib/*" -d %JAR_HOME%/TransCoder %JAR_HOME%/../*.java
-    copy %JAR_HOME%\MANIFEST.MF1 %JAR_HOME%\MANIFEST.MF /Y
-    pause
-    "%JAVA_HOME%jar.exe" cvfm %JAR_HOME%/FileManagerUI.jar %JAR_HOME%/MANIFEST.MF -C %JAR_HOME%/TransCoder/ .
-    pause
+    echo 1/6. prepare MF file
+    "%JAVA_HOME%javac.exe" -encoding UTF-8 -cp "..\lib\*" -d %JAR_HOME%\TransCoder %JAR_HOME%\..\*.java
+    copy %JAR_HOME%\MANIFEST.MF3 %JAR_HOME%\MANIFEST.MF /Y
+
+    echo 2/6. create temp files
+    if exist temp-jar-content rmdir /s /q temp-jar-content
+    mkdir temp-jar-content
+
+    echo 3/6. copy java files to temp directoty
+    xcopy "%JAR_HOME%\..\*.java" temp-jar-content /Y
+
+    echo 4/6. copy class files to temp directoty
+    xcopy "%JAR_HOME%\TransCoder\*" temp-jar-content /S /Y
+
+    echo 5/6. package jar with sources
+    "%JAVA_HOME%jar.exe" cvfm  %JAR_HOME%/FileManagerUI.jar %JAR_HOME%/MANIFEST.MF -C temp-jar-content/ .
+
+    "%JAVA_HOME%java.exe" -version
+    pause 
+    echo 6/6. clean up temp files
+    rmdir /S /Q temp-jar-content
     del /Q %JAR_HOME%\TransCoder\*.*
     del /Q %JAR_HOME%\MANIFEST.MF
 ) else (
